+++ src/sagas/queue.saga.test.js
@@ -6,8 +6,12 @@ import rootSaga, {
 import { getNextActionForProjectId } from '../reducers/queue.reducer';
 import {
   startNextActionInQueue,
+  installDependenciesStart,
+  uninstallDependenciesStart,
+  INSTALL_DEPENDENCIES_START,
   INSTALL_DEPENDENCIES_ERROR,
   INSTALL_DEPENDENCIES_FINISH,
+  UNINSTALL_DEPENDENCIES_START,
   UNINSTALL_DEPENDENCIES_ERROR,
   UNINSTALL_DEPENDENCIES_FINISH,
   START_NEXT_ACTION_IN_QUEUE,
@@ -40,6 +44,50 @@ describe('handleQueueActionCompleted saga', () => {
       select(getNextActionForProjectId, projectId)
     );
     expect(saga.next().done).toBe(true);
+  });
+});
+
+describe('handleStartNextActionInQueue saga', () => {
+  const projectId = 'foo';
+
+  let saga;
+  beforeEach(() => {
+    saga = handleStartNextActionInQueue({ projectId });
+  });
+
+  it('should do nothing if the queue is empty', () => {
+    expect(saga.next().value).toEqual(
+      select(getNextActionForProjectId, projectId)
+    );
+    expect(saga.next().done).toBe(true);
+  });
+
+  it(`should dispatch ${INSTALL_DEPENDENCIES_START} if an install action is queued`, () => {
+    const nextAction = {
+      action: 'install',
+      dependencies: [{ name: 'redux' }],
+    };
+
+    expect(saga.next().value).toEqual(
+      select(getNextActionForProjectId, projectId)
+    );
+    expect(saga.next(nextAction).value).toEqual(
+      put(installDependenciesStart(projectId, nextAction.dependencies))
+    );
+  });
+
+  it(`should dispatch ${UNINSTALL_DEPENDENCIES_START} if an uninstall action is queued`, () => {
+    const nextAction = {
+      action: 'uninstall',
+      dependencies: [{ name: 'redux' }],
+    };
+
+    expect(saga.next().value).toEqual(
+      select(getNextActionForProjectId, projectId)
+    );
+    expect(saga.next(nextAction).value).toEqual(
+      put(uninstallDependenciesStart(projectId, nextAction.dependencies))
+    );
   });
 });
